requirements.txt
python-telegram-bot==20.7
python-binance==1.0.19
python-dotenv==1.0.0
